+++ deltafi-ui/src/store/modules/errors/action-types.ts
@@ -1,0 +1,3 @@
+export enum ErrorsActionTypes {
+  FETCH_ERROR_COUNT = 'FETCH_ERROR_COUNT'
+}