+++ deltafi-ui/src/store/modules/errors/mutation-types.ts
@@ -1,0 +1,3 @@
+export enum ErrorsMutationTypes {
+  SET_ERROR_COUNT = 'SET_ERROR_COUNT',
+}