+++ deltafi-ui/src/utils/UtilFunctions.ts
@@ -36,4 +36,10 @@ export class UtilFunctions {
       return days.toFixed(precision) + "d";
     }
   }
+
+  pluralize(count: number, singular:string, plural?:string) {
+    if (count === 1) return `${count} ${singular}`;
+    if (plural) return `${count} ${plural}`;
+    return `${count} ${singular}s`;
+  }
 }