+++ deltafi-ui/CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Changed
+- Errors can now be acknowledged
+- Added Show/Hide Acknowledged button to Errors page
+- Date range selected has been removed from Errors page
+
 ## [0.16.3] - 2022-01-19
 ### Added
 - Persisted Search Options on DeltaFile Search page